class-powered socket used
Socket is now a class with the power of it. It holds the socketId
inside, so we can juste use write method without to know the id.
The read line and int8 is moved into the socket (the socket read the
data, not the httpparser, which is used just for http !)
in the httpServer subscript, I re-change set (newValue) : why is that
???, newValue is the default name, so you don’t have to name it that
way, except if you want a different name

I use deinit to release the socket. I test with Instrument with deinit
and release, and I don’t see any difference, but you can change if you
want, the release is still in comment.